code review: split authenticator in multiple files
